[mujoco] support ros melodic

+++ aerial_robot_simulation/scripts/convert.py
@@ -13,20 +13,31 @@ if len(meshdir) != 6:
 meshdir = meshdir[5]
 
 def delete_all():
-    for col in bpy.data.collections:
-        for item in col.objects:
-            col.objects.unlink(item)
+    if bpy.app.version[1] > 80:
+        for col in bpy.data.collections:
+            for item in col.objects:
+                col.objects.unlink(item)
+                bpy.data.objects.remove(item, do_unlink=True)
+
+        for item in bpy.context.scene.collection.objects:
+            bpy.context.scene.collection.objects.unlink(item)
             bpy.data.objects.remove(item, do_unlink=True)
 
-    for item in bpy.context.scene.collection.objects:
-        bpy.context.scene.collection.objects.unlink(item)
-        bpy.data.objects.remove(item, do_unlink=True)
+        for item in bpy.data.meshes:
+            bpy.data.meshes.remove(item, do_unlink=True)
 
-    for item in bpy.data.meshes:
-        bpy.data.meshes.remove(item, do_unlink=True)
+        # for item in bpy.data.materials:
+        #     bpy.data.materials.remove(item)
 
-    # for item in bpy.data.materials:
-    #     bpy.data.materials.remove(item)
+    if bpy.app.version[1] <= 79:
+        for item in bpy.context.scene.objects:
+            bpy.context.scene.objects.unlink(item)
+        for item in bpy.data.objects:
+            bpy.data.objects.remove(item)
+        for item in bpy.data.meshes:
+            bpy.data.meshes.remove(item)
+        for item in bpy.data.materials:
+            bpy.data.materials.remove(item)
 
 def process_subdirectories(root):
     for foldername, subfolders, filenames in os.walk(root):
